+++ client/main.py
@@ -1,0 +1,6 @@
+def Main() -> None:
+    pass
+
+
+if __name__ == "__main__":
+    Main()